Cleanup TreeWidget component
