+++ apps/test-app/frontend/src/components/tree-widget/TreeWidget.css
@@ -3,19 +3,12 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 
-.tree-widget {
-  display: flex;
-  flex-direction: column;
-}
-
 .tree-widget-tabs-content {
   padding: unset;
 }
 
 .tree-widget-header {
-  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.25);
   padding: 0.5em;
-  margin-bottom: 5px;
   display: flex;
   gap: 0.5em;
 }
