+++ apps/test-app/frontend/src/components/tree-widget/rules-driven/Tree.tsx
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import "../TreeWidget.css";
-import { useCallback, useEffect } from "react";
-import { SelectionMode } from "@itwin/components-react";
+import { useCallback, useEffect, useState } from "react";
+import { FilteringInputStatus, SelectionMode } from "@itwin/components-react";
 import { IModelConnection } from "@itwin/core-frontend";
 import {
   DiagnosticsProps,
@@ -15,8 +15,65 @@ import {
   UnifiedSelectionTreeEventHandler,
   usePresentationTreeState,
 } from "@itwin/presentation-components";
+import { TreeWidgetHeader, useTreeHeight } from "../TreeHeader";
 
 const PAGING_SIZE = 10;
+
+interface RulesDrivenTreeWidgetProps {
+  imodel: IModelConnection;
+  rulesetId?: string;
+  height?: number;
+  width?: number;
+}
+
+export function RulesDrivenTreeWidget(props: RulesDrivenTreeWidgetProps) {
+  const { rulesetId, imodel, width, height } = props;
+  const [diagnosticsOptions, setDiagnosticsOptions] = useState<DiagnosticsProps>({ ruleDiagnostics: undefined, devDiagnostics: undefined });
+  const [filter, setFilter] = useState("");
+  const [filteringStatus, setFilteringStatus] = useState(FilteringInputStatus.ReadyToFilter);
+  const [matchesCount, setMatchesCount] = useState<number>();
+  const [activeMatchIndex, setActiveMatchIndex] = useState(0);
+
+  const { headerRef, treeHeight } = useTreeHeight(height);
+  const onFilteringStateChange = useCallback((isFiltering: boolean, newMatchesCount: number | undefined) => {
+    setFilteringStatus(
+      isFiltering
+        ? FilteringInputStatus.FilteringInProgress
+        : undefined !== newMatchesCount
+          ? FilteringInputStatus.FilteringFinished
+          : FilteringInputStatus.ReadyToFilter,
+    );
+    setMatchesCount(newMatchesCount);
+  }, []);
+  return (
+    <>
+      <TreeWidgetHeader
+        ref={headerRef}
+        onFilterChange={setFilter}
+        filteringStatus={filteringStatus}
+        showFilteringInput={!!rulesetId}
+        onActiveMatchIndexChange={setActiveMatchIndex}
+        matchesCount={matchesCount}
+        onDiagnosticsOptionsChange={setDiagnosticsOptions}
+      />
+      <div className="filtered-tree">
+        {rulesetId && width && treeHeight ? (
+          <>
+            <Tree
+              imodel={imodel}
+              rulesetId={rulesetId}
+              diagnostics={diagnosticsOptions}
+              filtering={{ filter, activeMatchIndex, onFilteringStateChange }}
+              width={width}
+              height={treeHeight}
+            />
+            {filteringStatus === FilteringInputStatus.FilteringInProgress ? <div className="filtered-tree-overlay" /> : null}
+          </>
+        ) : null}
+      </div>
+    </>
+  );
+}
 
 interface Props {
   imodel: IModelConnection;
@@ -31,7 +88,7 @@ interface Props {
   height: number;
 }
 
-export function Tree(props: Props) {
+function Tree(props: Props) {
   const { filter, onFilteringStateChange, activeMatchIndex } = props.filtering;
   const state = usePresentationTreeState({
     imodel: props.imodel,
